Various renames and ignoring compilation units
