+++ src/net/client/mod.rs
@@ -4,8 +4,6 @@ pub use error::Error;
 
 use crate::{
     helpers::{
-        query::{PrepareQuery, QueryConfig, QueryInput},
-        transport::ByteArrStream,
         HelperIdentity,
     },
     net::{discovery::peer, http_serde},
@@ -187,14 +185,15 @@ impl MpcHelperClient {
     }
 }
 
+
+#[cfg(never)]
 #[cfg(all(test, not(feature = "shuttle")))]
 mod tests {
     use super::*;
     use crate::{
         ff::{FieldType, Fp31},
         helpers::{
-            query::{QueryCommand, QueryType},
-            CommandEnvelope, CommandOrigin, RoleAssignment, TransportCommand,
+            transport::query::QueryType,
             MESSAGE_PAYLOAD_SIZE_BYTES,
         },
         net::{server::BindTarget, MpcHelperServer},
