+++ src/net/server/handlers/query/step.rs
@@ -1,5 +1,4 @@
 use crate::{
-    helpers::{transport::TransportCommand, CommandEnvelope, CommandOrigin},
     net::{http_serde, server::Error},
     protocol::QueryId,
     sync::{Arc, Mutex},
