+++ src/net/transport.rs
@@ -1,8 +1,7 @@
 use crate::{
     helpers::{
         query::QueryCommand,
-        transport::{SubscriptionType, Transport, TransportCommand, TransportError},
-        CommandEnvelope, HelperIdentity,
+        HelperIdentity,
     },
     net::{
         client::MpcHelperClient,
