+++ src/query/mod.rs
@@ -4,4 +4,5 @@ mod state;
 
 pub use executor::Result as ProtocolResult;
 
+#[cfg(never)]
 pub use processor::{NewQueryError, Processor};
