+++ src/test_fixture/transport/network.rs
@@ -1,29 +1,30 @@
 use crate::{
-    helpers::{HelperIdentity, Transport},
+    helpers::{HelperIdentity},
     sync::{Arc, Weak},
-    test_fixture::transport::InMemoryTransport,
 };
+use crate::helpers::transport::ChannelledTransport;
+use crate::test_fixture::transport::transport::TransportCallbacks;
+use crate::test_fixture::transport::InMemoryChannelledTransport;
 
 /// Container for all active transports
 #[derive(Clone)]
 pub struct InMemoryNetwork {
-    pub transports: [Arc<InMemoryTransport>; 3],
+    pub transports: [Arc<InMemoryChannelledTransport>; 3],
 }
 
 impl Default for InMemoryNetwork {
     fn default() -> Self {
         let [mut first, mut second, mut third] = [
-            InMemoryTransport::setup(1.try_into().unwrap()),
-            InMemoryTransport::setup(2.try_into().unwrap()),
-            InMemoryTransport::setup(3.try_into().unwrap()),
+            InMemoryChannelledTransport::with_stub_callbacks(1.try_into().unwrap()),
+            InMemoryChannelledTransport::with_stub_callbacks(2.try_into().unwrap()),
+            InMemoryChannelledTransport::with_stub_callbacks(3.try_into().unwrap()),
         ];
-
-        InMemoryTransport::link(&mut first, &mut second);
-        InMemoryTransport::link(&mut second, &mut third);
-        InMemoryTransport::link(&mut third, &mut first);
+        first.connect(&mut second);
+        second.connect(&mut third);
+        third.connect(&mut first);
 
         Self {
-            transports: [first.listen(), second.listen(), third.listen()].map(Arc::new),
+            transports: [first.start(), second.start(), third.start()]
         }
     }
 }
@@ -41,7 +42,7 @@ impl InMemoryNetwork {
     }
 
     #[must_use]
-    pub fn transport(&self, id: HelperIdentity) -> Option<impl Transport + Clone> {
+    pub fn transport(&self, id: HelperIdentity) -> Option<impl ChannelledTransport> {
         self.transports
             .iter()
             .find(|t| t.identity() == id)
@@ -50,13 +51,14 @@ impl InMemoryNetwork {
 
     #[allow(clippy::missing_panics_doc)]
     #[must_use]
-    pub fn transports(&self) -> [impl Transport + Clone; 3] {
-        let transports: [Weak<InMemoryTransport>; 3] = self
+    pub fn transports(&self) -> [impl ChannelledTransport + Clone; 3] {
+        let transports: [Weak<InMemoryChannelledTransport>; 3] = self
             .transports
             .iter()
             .map(Arc::downgrade)
             .collect::<Vec<_>>()
             .try_into()
+            .map_err(|_| "What is dead may never die")
             .unwrap();
         transports
     }
